chore: rename bottomSheet component to free up namespace
